requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
Django==4.0.2
django-autoslug==1.9.8
django-bootstrap-form==3.4
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
future==0.18.2
html5lib==1.1
Pillow==9.0.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2022.1
reportlab==3.6.7
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
webencodings==0.5.1
xhtml2pdf==0.2.5
